move inc/utils to TermApp
