+++ src/cascadia/TerminalSettingsModel/Command.idl
@@ -17,7 +17,7 @@ namespace Microsoft.Terminal.Settings.Model
         String KeyChordText;
 
         Windows.UI.Xaml.Controls.IconSource IconSource;
-        void RefreshIcon();
+        String IconPath { get; };
 
         Boolean HasNestedCommands { get; };
         Windows.Foundation.Collections.IMapView<String, Command> NestedCommands { get; };
